% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovAhat_Sigmahat_Gamma.R
\name{CovAhat_Sigmahat_Gamma}
\alias{CovAhat_Sigmahat_Gamma}
\title{CovAhat_Sigmahat_Gamma}
\usage{
CovAhat_Sigmahat_Gamma(p, X, Z, eta, lags)
}
\arguments{
\item{p}{Number of lags in the VAR model}

\item{X}{VAR "right-hand" variables}

\item{Z}{external instrument}

\item{eta}{eta}

\item{lags}{Newey-West lags}
}
\value{
WHataux asymptotic variance of

WHat asymptotic variance of

V Matrix such that
}
\description{
Estimates the asymptotic covariance matrix
}
