% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gmatrices.R
\name{Gmatrices}
\alias{Gmatrices}
\title{Gmatrices}
\usage{
Gmatrices(AL, C, p, hori, n)
}
\arguments{
\item{AL}{VAR model coefficients}

\item{C}{MA representation coefficients}

\item{p}{lag order}

\item{hori}{forecast horizon}

\item{n}{number of variables}
}
\value{
G: derivatives of C wrt A
}
\description{
Computes the derivatives of vec(C) wrt vec(A) based on Lütkepohl H. New introduction to multiple time series analysis. Springer, 2007.
}
\seealso{
https://pdfs.semanticscholar.org/3e18/3a5ec97ff636363e4deedff7eaeee9d894c9.pdf
}
