% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARep.R
\name{MARep}
\alias{MARep}
\title{MARep}
\usage{
MARep(AL, p, hori)
}
\arguments{
\item{AL}{VAR model coefficients}

\item{p}{lag order}

\item{hori}{forecast horizon}
}
\value{
C MA representation coefficients
}
\description{
Transforms the A(L) parameters of a reduced-form VAR into the coefficients C of the MA representation.
}
