% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSWfunction.R
\name{MSWfunction}
\alias{MSWfunction}
\title{MARep}
\usage{
MSWfunction(confidence, nvar, scale, horizons, RForm)
}
\arguments{
\item{confidence}{confidence level}

\item{nvar}{variable defining the normalization}

\item{scale}{scale of the shock}

\item{horizons}{Number of horizons for the IRF}

\item{RForm}{reduced-form structure}

\item{display_on}{dummy variable}
}
\value{
InferenceMSW: Structure containing the MSW weak-iv robust confidence interval

Plugin: Structure containing standard plug-in inference

Chol: Cholesky IRFs
}
\description{
Reports the confidence interval for IRF coefficients described in Montiel-Olea, Stock, and Watson (2017).
}
\seealso{
https://github.com/jm4474/SVARIV

https://uc4384f22718973d3d912d614f2e.dl.dropboxusercontent.com/cd/0/inline2/Aslwi_FA9z6v37dfxMeUP3mEaP-a4ZA8RFrXd5h_EymM1lqGy-UUwGJB3TxqRxluMhE9ebnvHEMZktfcUmJ1v4mGSxeS_ZEDJBCrjGvyopAnP44ClPLJ6kmM36F557kHnqKQjn___0XTNvPaWvXnkEC0Btt3MpVhn41q0ddL4qtZ1g9l2ooeVweHH9X5xCAaG3R28lpzivQzj8AZRVSPHs98yseh0L9I10-G0frVqteUkjuQJQNAmFlEPpqegmDhpPo_vaMScx1w1ivQ--mxoZli2uvBI1Pnen-nGIyD3qT_oT-0r3K6bjwk1vI71fgMtGw6ZhWWFuWv48Oq-vg7lN1fKgPEyvcJMcdIISnra0z-1w/file
}
