% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oil.R
\docType{data}
\name{oil}
\alias{oil}
\title{Macro data on oil shocks}
\format{
Matrix with 6 rows (see examples)
}
\source{
https://github.com/jm4474/SVARIV
}
\usage{
data(oil)
}
\description{
This folder contains data for the oil-SVAR used in the paper as empirical illustration, as well as a the fiscal-SVAR application in “Marginal Tax Rates and Income: New Time Series Evidence”, Mertens and Montiel Olea (2018); Forthcoming at The Quarterly Journal of Economics
}
\examples{
data(oil)
ydata<-data[,3:5]
z<-data[,6]
years<-data[,1:2]

}
\references{
Olea, Stock and Watson (2018)
}
\keyword{datasets}
