% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_irf.R
\name{pretty.irf}
\alias{pretty.irf}
\title{CovAhat_Sigmahat_Gamma}
\usage{
\method{pretty}{irf}(
  data,
  shock_names,
  pretty_names = NULL,
  cum = F,
  confidence_type = "msw",
  manual_color = NULL,
  legend = F,
  title = NULL,
  same_scale = T,
  shock_sign = "positive"
)
}
\arguments{
\item{data}{list of var$irf objects from SVARIV}

\item{shock_names}{instrument names}

\item{pretty_names}{vector of pretty names for y-axis}

\item{cum}{logical. Cumulative? BEWARE: Possible mistake here}

\item{confidence_type}{vector of confidence types}

\item{manual_color}{specify colors per Instrument. Input is a named vecor with shock_names and the corresponding color (hex)}

\item{legend}{logical. inculde or exclude legends}

\item{title}{vector of headings for each column}

\item{same_scale}{logical value. Should all columns use the same scale?}
}
\value{
ggplot wrapper
}
\description{
Estimates the asymptotic covariance matrix
}
\examples{
p           = 24     #Number of lags in the VAR model
confidence  = confidence=c(0.6,0.9,0.95)    #Confidence Level for the standard and weak-IV robust confidence set
NWlags      = 0;  # Newey-West lags(set it to 0 to compute heteroskedasticity robust std errors)
norm        = 1; # Variable used for normalization
scale       = 1; # Scale of the shock
horizons    = 20; #Number of horizons for the Impulse Response Functions(IRFs)
colnames(oil)<-c("a","b","c","d","year","month")
ydata<-oil[,1:3]
z<-oil[,4]

VAR<-SVARIV(ydata,z,p,confidence,NWlags,norm,scale,horizons,instrument_name="test")
sh.col<-      c("#E41A1C")
names(sh.col)<-c("test")
pretty.irf(data=list(VAR$irfs),shock_names="test",pretty_names=c("a","b","c"),manual_color=sh.col,title="subheading")
}
