% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RForm_VAR.R
\name{RForm_VAR}
\alias{RForm_VAR}
\title{RForm_VAR}
\usage{
RForm_VAR(TSL, p, W = NULL)
}
\arguments{
\item{TSL}{matrix of time series}

\item{p}{number of lags in the VAR model}

\item{W}{Matrix of exogenous regressors}
}
\value{
AL Least-squares estimator of the VAR coefficients

Sigma Least-squares estimator of the VAR residuals

eta VAR model residuals

X Matrix of VAR covariates

Y VAR matrix of endogenous regressors

If no exogenous regressors are specified, our estimation always includes a constant.
}
\description{
Provides reduced form estimators of a VAR(p) model
}
\seealso{
\href{https://github.com/jm4474/SVARIV}

\href{https://uc4384f22718973d3d912d614f2e.dl.dropboxusercontent.com/cd/0/inline2/Aslwi_FA9z6v37dfxMeUP3mEaP-a4ZA8RFrXd5h_EymM1lqGy-UUwGJB3TxqRxluMhE9ebnvHEMZktfcUmJ1v4mGSxeS_ZEDJBCrjGvyopAnP44ClPLJ6kmM36F557kHnqKQjn___0XTNvPaWvXnkEC0Btt3MpVhn41q0ddL4qtZ1g9l2ooeVweHH9X5xCAaG3R28lpzivQzj8AZRVSPHs98yseh0L9I10-G0frVqteUkjuQJQNAmFlEPpqegmDhpPo_vaMScx1w1ivQ--mxoZli2uvBI1Pnen-nGIyD3qT_oT-0r3K6bjwk1vI71fgMtGw6ZhWWFuWv48Oq-vg7lN1fKgPEyvcJMcdIISnra0z-1w/file}
}
