% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVARIV.R
\name{SVARIV}
\alias{SVARIV}
\title{SvAR-IV.}
\usage{
SVARIV(
  ydata,
  z,
  p,
  confidence,
  NWlags,
  norm,
  scale,
  horizons,
  ci_type = c("msw"),
  print_wald = T,
  instrument_name
)
}
\arguments{
\item{ydata}{Endogenous variables from the VAR model}

\item{z}{External instrumental variable}

\item{p}{Number of lags in the VAR model}

\item{confidence}{Value for the standard and weak-IV robust confidence set}

\item{NWlags}{Newey-West lags (set it to 0 to compute heteroskedasticity robust std errors)}

\item{norm}{Variable used for normalization}

\item{scale}{Scale of the shock}

\item{horizons}{Number of horizons for the Impulse Response Functions (does not include the impact or horizon 0)}

\item{ci_type}{confidence intervals to include (choose from "msw" (Montiel, Stock and Watson, 2020), "delta" or "plugin"), just triggers msw waldtest. specify ci by prett.irf command}

\item{print_wald}{Number of horizons for the Impulse Response Functions (does not include the impact or horizon 0)}

\item{instrument_name}{Number of horizons for the Impulse Response Functions (does not include the impact or horizon 0)}
}
\value{
irfs:       list containing all irf data

waldstat: contain msw waldstat
}
\description{
Implements standard and weak-IV robust SVAR-IV inference.
}
\examples{



p           = 24    #Number of lags in the VAR model
NWlags      = 0;  # Newey-West lags(if it is neccessary to account for time series autocorrelation)
norm        = 1; # Variable used for normalization
scale       = 1; # Scale of the shock
horizons    = 20; #Number of horizons for the Impulse Response Functions(IRFs)
confidence=c(0.6,0.9,0.95);
data(oil)
colnames(oil)<-c("a","b","c","d","year","month")
ydata<-oil[,1:3]
z<-oil[,4]
VAR<-SVARIV(ydata,z,p,confidence,NWlags,norm,scale,horizons,instrument_name="test")

}
