% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVARIV_Check.R
\name{SVARIV_Check}
\alias{SVARIV_Check}
\title{SVARIV_Check}
\usage{
SVARIV_Check(p, confidence, ydata, z, NWlags, norm, scale, horizons)
}
\arguments{
\item{p}{Number of lags in the VAR model}

\item{confidence}{Value for the standard and weak-IV robust confidence set}

\item{ydata}{Endogenous variables from the VAR model}

\item{z}{External instrumental variable}

\item{NWlags}{Newey-West lags}

\item{norm}{Variable used for normalization}

\item{scale}{Scale of the shock}

\item{horizons}{Number of horizons for the Impulse Response Functions}
}
\value{
None
}
\description{
Checks whether the inputs from SVARIV are valid.
}
