% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oil.R
\docType{data}
\name{oil}
\alias{oil}
\title{Macro data on oil shocks}
\format{
Matrix with 6 rows (see examples)
}
\source{
\href{https://github.com/jm4474/SVARIV}
}
\usage{
data(oil)
}
\description{
This folder contains data for the oil-SVAR used in the paper as empirical illustration, as well as a the fiscal-SVAR application in “Marginal Tax Rates and Income: New Time Series Evidence”, Mertens and Montiel Olea (2018); Forthcoming at The Quarterly Journal of Economics
}
\examples{
data(oil)
ydata<-data[,3:5]
z<-data[,6]
years<-data[,1:2]

}
\references{
Olea, Stock and Watson (2018)
(\href{https://uc4384f22718973d3d912d614f2e.dl.dropboxusercontent.com/cd/0/inline2/Aslwi_FA9z6v37dfxMeUP3mEaP-a4ZA8RFrXd5h_EymM1lqGy-UUwGJB3TxqRxluMhE9ebnvHEMZktfcUmJ1v4mGSxeS_ZEDJBCrjGvyopAnP44ClPLJ6kmM36F557kHnqKQjn___0XTNvPaWvXnkEC0Btt3MpVhn41q0ddL4qtZ1g9l2ooeVweHH9X5xCAaG3R28lpzivQzj8AZRVSPHs98yseh0L9I10-G0frVqteUkjuQJQNAmFlEPpqegmDhpPo_vaMScx1w1ivQ--mxoZli2uvBI1Pnen-nGIyD3qT_oT-0r3K6bjwk1vI71fgMtGw6ZhWWFuWv48Oq-vg7lN1fKgPEyvcJMcdIISnra0z-1w/file}{link})
}
\keyword{datasets}
