% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RForm_VAR.R
\name{RForm_VAR}
\alias{RForm_VAR}
\title{RForm_VAR}
\usage{
RForm_VAR(TSL, p, W = NULL)
}
\arguments{
\item{TSL}{matrix of time series}

\item{p}{number of lags in the VAR model}

\item{W}{Matrix of exogenous regressors}
}
\value{
AL Least-squares estimator of the VAR coefficients

Sigma Least-squares estimator of the VAR residuals

eta VAR model residuals

X Matrix of VAR covariates

Y VAR matrix of endogenous regressors

If no exogenous regressors are specified, our estimation always includes a constant.
}
\description{
Provides reduced form estimators of a VAR(p) model
}
\seealso{
https://github.com/jm4474/SVARIV
}
