% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVARIV.R
\name{SVARIV}
\alias{SVARIV}
\title{SvAR-IV.}
\usage{
SVARIV(ydata, z, p, confidence, NWlags, norm, scale, horizons)
}
\arguments{
\item{ydata}{Endogenous variables from the VAR model}

\item{z}{External instrumental variable}

\item{p}{Number of lags in the VAR model}

\item{confidence}{Value for the standard and weak-IV robust confidence set}

\item{NWlags}{Newey-West lags}

\item{norm}{Variable used for normalization}

\item{scale}{Scale of the shock}

\item{horizons}{Number of horizons for the Impulse Response Functions}
}
\value{
PLugin:       Structure containing standard plug-in inference

InferenceMSW: Structure containing the MSW weak-iv robust confidence interval

Chol:         Cholesky IRFs

RForm:        Structure containing the reduced form parameters
}
\description{
Implements standard and weak-IV robust SVAR-IV inference.
}
