% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVARIV_Check.R
\name{SVARIV_Check}
\alias{SVARIV_Check}
\title{SVARIV_Check}
\usage{
SVARIV_Check(
  p,
  confidence,
  ydata,
  z,
  NWlags,
  norm,
  scale,
  horizons,
  savdir,
  columnnames,
  IRFselect,
  cumselect,
  time,
  dataset_name
)
}
\arguments{
\item{p}{Number of lags in the VAR model}

\item{confidence}{Value for the standard and weak-IV robust confidence set}

\item{ydata}{Endogenous variables from the VAR model}

\item{z}{External instrumental variable}

\item{NWlags}{Newey-West lags}

\item{norm}{Variable used for normalization}

\item{scale}{Scale of the shock}

\item{horizons}{Number of horizons for the Impulse Response Functions}

\item{savdir}{Directory where the figures generated will be saved}

\item{columnnames}{Vector with the names for the endogenous variables, in the same order as ydata}

\item{IRFselect}{Indices for the variables that the user wants separate IRF plots for}

\item{cumselect}{Indices for the variables that the user wants cumulative IRF plots for}

\item{time}{Time unit for the dataset e.g. year, month, etc.}

\item{dataset_name}{The name of the dataset used for generating the figures}
}
\value{
None
}
\description{
Checks whether the inputs from SVARIV are valid.
}
